resolved bug
